Refactor admin modules
