+++ src/meshapi/admin/__init__.py
@@ -1,11 +1,3 @@
-from .access_point import *
 from .admin import *
-from .building import *
-from .device import *
 from .inlines import *
-from .install import *
-from .link import *
-from .los import *
-from .member import *
-from .node import *
-from .sector import *
+from .models import *
